Add Shipper_Territory

+++ Northwind.API/Models/Order.cs
@@ -17,7 +17,7 @@ namespace Northwind.API.Models
         public DateTime? RequiredDate { get; set; }
         public DateTime? ShippedDate { get; set; }
         public int ShipVia { get; set; }
-        public decimal Freight { get; set; }
+        public double Freight { get; set; }
         public string ShipName { get; set; }
         public string ShipAddress { get; set; }
         public string ShipCity { get; set; }
